LU-9758 build: allow disabling lustre test and iokit rpm creation
While attempting to create a basic set of rpms that didn't
include the lustre test and iokit rpm I encountered build
breakage. Test the lustre_test conditional in the spec file
so we don't attempt to build special lustre test rpms. The
lustre test rpm is actually dependant on the lustre iokit
rpm so if --disable-iokit is set we should disable lustre
test rpms generation as well.

Test-Parameters: trivial

Change-Id: I0adeb0eded1a3b99481fc4b7ad3cda2afd32b6d2
Signed-off-by: James Simmons <uja.ornl@yahoo.com>
Reviewed-on: https://review.whamcloud.com/27985
Reviewed-by: Bob Glossman <bob.glossman@intel.com>
Tested-by: Jenkins
Tested-by: Maloo <hpdd-maloo@intel.com>
Reviewed-by: Dmitry Eremin <dmitry.eremin@intel.com>
Reviewed-by: Oleg Drokin <oleg.drokin@intel.com>