LU-7356 build: handle configure options for mount plugin
Currently if you build lustre with configure options
--disable-shared and/or --disable-static it will fail
to build in the case of the mount plugins. The mount
plugins are always built to be sharable code using
libtool but this is not compatiable with the confiure
--disable-share option. The solution was to create a
temporary static library and for the static case we
directly link into the lustre utilites. If shared
libraries are enabled then that temporary library
is turned into a dynamic plugin. The lustre utilites
was updated to handle both cases.

Test-Parameters: trivial

Change-Id: I0eec25c8bce71c0e5c533ff3dfbe752677f37a26
Signed-off-by: James Simmons <uja.ornl@yahoo.com>
Signed-off-by: Nathaniel Clark <nathaniel.l.clark@intel.com>
Reviewed-on: https://review.whamcloud.com/17581
Tested-by: Jenkins
Tested-by: Maloo <hpdd-maloo@intel.com>
Reviewed-by: Dmitry Eremin <dmitry.eremin@intel.com>
Reviewed-by: Bob Glossman <bob.glossman@intel.com>
Reviewed-by: Oleg Drokin <oleg.drokin@intel.com>