LU-8602 gss: Properly port gss to newer crypto api.
In newer kernels the old crypto_hash_* was removed which
requires the gss layer to now user crypto_ahash_*. Pieces
of the port were done incorrectly which this patch addresses.
In the process cleanup some cases where data was leaked
in the case of the sg_tables.

Address the limitation of the gss keys in that they are hard
coded to u16 variables that are not guaranteed to be constant
and it will prove to be challenge in the future to add new crypto
algorithms.
Keep u16 representation of hash and crypto algorithms in Shared-Keys
as they are stored on disk, to not break compatibility with already
existing keys. But pass in to kernel space the name string
of the crypto algorithm we want to use.

Also deal with timer_setup and key.usage new format introduced in
newer kernels.

Test-Parameters: testlist=sanity-sec envdefinitions=SHARED_KEY=true
Change-Id: I90ef58104a1955ce12603173964a6878f60b601e
Signed-off-by: James Simmons <uja.ornl@yahoo.com>
Signed-off-by: Sebastien Buisson <sbuisson@ddn.com>
Reviewed-on: https://review.whamcloud.com/28309
Reviewed-by: John L. Hammond <jhammond@whamcloud.com>
Tested-by: Jenkins
Tested-by: Maloo <hpdd-maloo@intel.com>
Reviewed-by: Oleg Drokin <green@whamcloud.com>