LU-8066 obd: fix LPROC_SEQ_FOPS macros for debugfs
While porting to debugfs I discovered a bug in that
for the debugfs case no data was being passed into
single_open(). This would cause a crash when reading
debugfs files since the seq_file private field was
NULL. Second change is the rename of the *WO_TYPE*
macro to LPROC_SEQ_FOPS_WR_ONLY to match upstream.

Change-Id: I9b6baebecaeb82795d7055148533862489473a1d
Signed-off-by: James Simmons <uja.ornl@yahoo.com>
Reviewed-on: https://review.whamcloud.com/26650
Tested-by: Jenkins
Tested-by: Maloo <hpdd-maloo@intel.com>
Reviewed-by: Emoly Liu <emoly.liu@intel.com>
Reviewed-by: Dmitry Eremin <dmitry.eremin@intel.com>
Reviewed-by: Oleg Drokin <oleg.drokin@intel.com>