LU-6210 utils: Use C99 struct initializers in lnetctl
This patch makes no functional changes.  The option struct
initializers in lnetctl are updated to C99 syntax.  The short and
long options are renamed to short_opts and long_opts for consistency.

C89 positional initializers require values to be placed in the
correct order. This will cause errors if the fields of the struct
definition are reordered or fields are added or removed. C99 named
initializers avoid this problem, and also automatically clear any
values that are not explicitly set.

Test-Parameters: trivial
Signed-off-by: Steve Guminski <stephenx.guminski@intel.com>
Change-Id: I1c1483a57aea918dce84afd0c7e94e31324c189e
Reviewed-on: https://review.whamcloud.com/28423
Tested-by: Jenkins
Tested-by: Maloo <hpdd-maloo@intel.com>
Reviewed-by: James Simmons <uja.ornl@yahoo.com>
Reviewed-by: John L. Hammond <john.hammond@intel.com>
Reviewed-by: Oleg Drokin <oleg.drokin@intel.com>