LU-9580 tests: remove performance-sanity tests from ALWAYS_EXCEPT
Remove performance-sanity tests 1 and 2 from ALWAYS_EXCEPT
list as well as tests test_1 and test_2 because all they
contain are calls to echo.
Tests 1 and 2 are associated with bugzilla ticket 15266 and
it is fixed. Yet, reviewing all comment in that ticket
reveals that tests 1 and 2 were never implemented.

Test-Parameters: trivial testlist=performance-sanity

Signed-off-by: dilip krishnagiri <dilipx.krishnagiri@intel.com>
Change-Id: I402474f9db0d1875bf9c4b5c071e9c27bd47ba28
Reviewed-on: https://review.whamcloud.com/27375
Reviewed-by: James Nunez <james.a.nunez@intel.com>
Tested-by: Jenkins
Tested-by: Maloo <hpdd-maloo@intel.com>
Reviewed-by: Saurabh Tandan <saurabh.tandan@intel.com>
Reviewed-by: Oleg Drokin <oleg.drokin@intel.com>