LU-6245 tests: remove it_test
The test application it_test pulls in lustre kernel code which is
the wrong approach and this will not work in the future when the
lustre utilites will be build against the upstream kernel lustre
code. Looking at where it_test is actually run I discovered it is
never used. Since this is the case we can delete it.

Test-Parameters: trivial

Signed-off-by: James Simmons <uja.ornl@yahoo.com>
Change-Id: I8fdcf279590b0edf1c8da6d9f4618964162ec762
Reviewed-on: https://review.whamcloud.com/28406
Tested-by: Jenkins
Tested-by: Maloo <hpdd-maloo@intel.com>
Reviewed-by: Dmitry Eremin <dmitry.eremin@intel.com>
Reviewed-by: John L. Hammond <john.hammond@intel.com>
Reviewed-by: Oleg Drokin <oleg.drokin@intel.com>