LUDOC-372 build: Fix a condition issue with the PDF build
This patch fixes an issue with the building of the PDF
version of the manual where sections designated to use
an "Introduced in Lustre 2.X" frame were not being
processed and the entire section heading was missing.

The technical issue that was occuring during the PDF
processing is that the template in the PDF stylesheet,
style/customstyle.fo, was dropping the content passed to it.
This did not effect non-section tags, however, for section tags,
this content included the section number and heading.  It is
resolved by capturing the content passed into the condition
template and relaying it onto the textdecoration_1 template
which makes the calls to render the decoration of the text.
This fix matches what the html template uses for rendering.

This patch does not affect the build for any other
versions of the manual as a different stylesheet is used
for each format.

Signed-off-by: Joseph Gmitter <joseph.gmitter@intel.com>
Change-Id: I7ccae3e6f74ea2501f84ea224b20a8bb56b85894
Reviewed-on: https://review.whamcloud.com/28583
Reviewed-by: Andreas Dilger <andreas.dilger@intel.com>
Tested-by: Jenkins