LU-8066 lnet: port lnet router to debugfs
Move all the lnet procfs variables to debugfs.

Linux-commit: b03f395a3e9bf4874fb58f4fe6033866d3b9f105

This brings the OpenSFS branch into sync with upstream
lustre client. Technically debugfs is the not the proper
place for stats but due to the upstream client being
uses for production systems we have to keep them :-(
New work will be done in the future to properly handle
stats using sysfs instead.

Test-Parameters: trivial envdefinitions=ONLY=215 testlist=sanity

Change-Id: Id9244d8525f0844d321f29af4a01e1fbce8e5884
Signed-off-by: Oleg Drokin <green@linuxhacker.ru>
Signed-off-by: Greg Kroah-Hartman <gregkh@linuxfoundation.org>
Signed-off-by: James Simmons <uja.ornl@yahoo.com>
Reviewed-on: https://review.whamcloud.com/26430
Reviewed-by: Olaf Weber <olaf.weber@hpe.com>
Tested-by: Jenkins
Tested-by: Maloo <hpdd-maloo@intel.com>
Reviewed-by: Dmitry Eremin <dmitry.eremin@intel.com>
Reviewed-by: Oleg Drokin <oleg.drokin@intel.com>