LU-9891 tests: Increase space not released for ZFS
Several Lustre tests calculate the free space on the
object storage servers. For servers running ZFS, the amount
of space released by ZFS is not 100% deterministic. Thus,
fs_log_size() will return the buffer size that we allow
the space to be off by. For ZFS, increase this buffer
from 400 to 512 KB.

Test-Parameters: trivial testgroup=review-zfs-part-2
Signed-off-by: James Nunez <james.a.nunez@intel.com>
Change-Id: I32e0ae3752d0ee0e9f0091ea779f8b53ba969a26
Reviewed-on: https://review.whamcloud.com/28682
Tested-by: Jenkins
Reviewed-by: Andreas Dilger <andreas.dilger@intel.com>
Reviewed-by: Patrick Farrell <paf@cray.com>
Tested-by: Maloo <hpdd-maloo@intel.com>