LU-9888 tests: Update disk2_7-zfs.tar.bz2 for quota
Add a blimit file with the larger 40960 limit in it, as this seems to
be how the image was created, but the default (without this file) is
20K.
Also add an ilimit file with 4 (default for new images).  Old default
value is 2.

This patch rolls in LU-9826 fix for larger ZFS
Lustre-change: https://review.whamcloud.com/28464
Lustre-commit: 6ab9866730c1bd2344869502586fc322f6ed7b71

This patch on master
Lustre-change: https://review.whamcloud.com/28820
Lustre-commit: f37b25fbb4acf3f96ebc61bb4d9137a8f0401e35

Signed-off-by: Nathaniel Clark <nathaniel.l.clark@intel.com>
Change-Id: I33921ac58a5252f3259145d5e00faedcd21559f9
Reviewed-by: Alex Zhuravlev <alexey.zhuravlev@intel.com>
Reviewed-by: Wei Liu <wei3.liu@intel.com>
Signed-off-by: Minh Diep <minh.diep@intel.com>
Reviewed-on: https://review.whamcloud.com/29068
Tested-by: Jenkins
Tested-by: Maloo <hpdd-maloo@intel.com>
Reviewed-by: John L. Hammond <john.hammond@intel.com>