LU-9785 lov: take lov layout lock for I/O with ignore_layout
A rule of thumb for taking lov layout configuration lock is if I/O
is initiated from LLITE layer, it should grab the lock. If an I/O
starts from the OSC layer, it won't be necessary because if the OSC
object exists, layout reconfiguration will move forward.

Right now CIT_MISC + ci_ignore_layout can identify the I/O from the
OSC layer, I just use this in lov_io_init() for this purpose. In the
future, an explicit bit may be defined for this.

Lustre-commit: e43b0e5c0ccbacd8adf30713babd865b5a7c58c7
Lustre-change: https://review.whamcloud.com/29638

Signed-off-by: Jinshan Xiong <jinshan.xiong@gmail.com>
Change-Id: I2fe37a957b5fb4161c4c723062f6469b915c1dd5
Reviewed-by: Andreas Dilger <andreas.dilger@intel.com>
Reviewed-by: Bobi Jam <bobijam@hotmail.com>
Reviewed-by: Fan Yong <fan.yong@intel.com>
Reviewed-by: Jinshan Xiong <jinshan.xiong@intel.com>
Reviewed-on: https://review.whamcloud.com/33965
Reviewed-by: Lai Siyao <lai.siyao@whamcloud.com>
Tested-by: Jenkins
Tested-by: Maloo <maloo@whamcloud.com>
Reviewed-by: Oleg Drokin <green@whamcloud.com>